fix bug in Parallel server. fix bug in DFS alg. Move ServerAbs class to Server file under the same namespace: server_side. Print solution path cost as double instead of int.

+++ MatrixSolution.cpp
@@ -63,5 +63,5 @@ std::string MatrixSolution::recursion_path(State<double> *state) {
   else {
     direction = UP;
   }
-  return recursion_path((*state).get_parent()) + direction + " (" + std::to_string((int)(*state).get_cost()) + ") ,";
+  return recursion_path((*state).get_parent()) + direction + " (" + std::to_string((double)(*state).get_cost()) + ") ,";
 }
